Remove client_id from non-PKCE access token requests;
As the client is already identified using the Authorization header in the request,
there is no need to pass the client_id parameter in the requests. The parameter
however is needed in the PCKE requests.

TODO: A test should be added to verify that the client_id parameter is
ignored if provided.